+++ backend/src/game/game.module.ts
@@ -11,6 +11,7 @@ import {ModeRepository} from 'src/record/mode/mode.repository';
 import {TypeRepository} from 'src/record/type/type.repository';
 import {JwtModule} from '@nestjs/jwt';
 import {JwtStrategy} from 'src/user/jwt.strategy';
+import { GameService } from './game.service';
 
 @Module({
   imports: [
@@ -30,6 +31,7 @@ import {JwtStrategy} from 'src/user/jwt.strategy';
     ModeRepository,
     TypeRepository,
     JwtStrategy,
+    GameService,
   ],
 })
 export class GameModule {}
