+++ backend/src/game/types/key_data.interface.ts
@@ -1,0 +1,4 @@
+export interface KeyData {
+  up: boolean;
+  down: boolean;
+}
