+++ backend/src/game/types/user_index.enum.ts
@@ -1,0 +1,4 @@
+export enum EUserIndex {
+  LEFT = 0,
+  RIGHT = 1,
+}
