+++ types/game/ball.interface.ts
@@ -1,7 +1,7 @@
 import { Coordinate } from "./coordinate.interface";
 import { Velocity } from "./velocity.interface";
 
-export const DEFAULT_BALL_SPEED = 3;
+export const DEFAULT_BALL_SPEED = 2;
 
 export interface Ball {
   pos: Coordinate;
