+++ types/game/game-info.interface.ts
@@ -2,6 +2,7 @@ import { Ball } from "./ball.interface";
 import { Coordinate } from "./coordinate.interface";
 
 export interface GameInfo {
+  timeStamp: number;
   leftPaddle: Coordinate;
   leftScore: number;
   rightPaddle: Coordinate;
